created notification diff util
